apps/weather-app/requirements.txt
reflex>=0.4.0
fastapi>=0.104.0
sqlalchemy>=2.0.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0